feat: implement unit tests in test_project.py
This was the hardest part. Had to read on unittest
and mocking since all my functions require direct
inputs and do not expect parameters. Was not direclty
covered in CS50P. Thank you youtube. After 3 hours,
finally managed to figure it out.... My project is done.
Thanks to CS50 and Prof. Malan for helping me discover
a new passion. If anyone reads this, thank you for taking
the time to check my project!